new base web search service
